Version 1.0
Did a complete rewrite of the script-code (this is the main reason for the version-number leap).
A separate config-file, which should be modified to reflect the user's preferences.
More variables for better customization.
Due to the rewrite, should now be more robust
Cursor movement and auto-start insert mode
Better control over the look of the comment templates

Read new installation and configuration instructions from the .vim files