Version 1.1
This is again a .vim file (and not a zip). The configuration file is pasted at the beginning of the script. Again, see the script
for instructions.

Can now choose between '@throws' and '@exception' tags.
When executed on single-line comments ("/** blah blah */") expands them into multiline comments preserving the text
Partial method documentation comment updating (adds tags instead of writing a new template)