Version 1.2
A bunch of new features (some seem to be actually useful ;-) :
Complete method comment updating: now adds/removes @return- and @param-tags form existing doc comments if no more applicable.
Added a new public function (+ mappings for it) for searching incomplete doc-comments (This is nice - a quick way to complete comments and to jump to next incomplete comment)
If executed on a line like "throw new RuntimeExeception()" will add an exception declaration to prev. javadoc comments.
Corrected a problem where updating didn't work if a new parameter's name began the same way than an existing one's.
Now behaves correctly when "extra stuff" (after the '{') is included in the range.
Added mapping for insert-mode to config-file.
Added an abbreviation (see config-file)
Now adds the method/class/if/while/for... name as an ending comment when executed on a line containing only whitespace and a '}'.
Comment generation can now be virtually fully automated by using an abbreviation for '{' (which is presented in the config-file).